Change system spec to use new design
